% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Data checker}
\usage{
check_data(x)
}
\arguments{
\item{x}{numeric values containing the data of direct relationship decision matrix.}
}
\value{
This function checks whether \code{data} is matrix or not. Returns a matrix and assign new names if not defined.
}
\description{
Throws an error message if data is not matrix format, checks names attribute of the matrix, assign new ones if not defined
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
