% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_criteria.R
\name{compare_criteria}
\alias{compare_criteria}
\title{Relation results}
\usage{
compare_criteria(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a \code{num} matrix.
}
\description{
Returns relation results that exceed threshold value of direct relationship decision matrix
}
\examples{
compare_criteria(dematel::hospitaldata)
compare_criteria(dematel::nurseselection)
compare_criteria(dematel::medicaldevice)
}
\seealso{
\link[base]{apply} function.
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
