% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dematel.R
\docType{package}
\name{dematel}
\alias{dematel}
\alias{dematel-package}
\title{A Technique of Multi-Criteria Decision Making}
\description{
Easily solve Multi-Criteria Decision Making Problems with Decision Making Trial and Evaluation Laboratory Technique.
}
\examples{
normalize_data(dematel::nurseselection)

total_relationship_matrix(dematel::nurseselection)

relationships_between_criteria(dematel::nurseselection)

visualize(dematel::nurseselection)

threshold_value(dematel::nurseselection)

compare_criteria(dematel::nurseselection)

}
\author{
Muhlis Ozdemir \email{muhlisoz@gmail.com} Yakup Celikbilek \email{yakupcelikbilek@gmail.com}
}
