% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_dematel.R
\name{execute_dematel}
\alias{execute_dematel}
\title{Complete Dematel Analysis}
\usage{
execute_dematel(x)
}
\arguments{
\item{x}{a matrix containing the values of normalized initial direct-relation decision matrix.}
}
\value{
This function executes all functions, conducts dematel analysis at once and returns
a \code{matrix} that contains data,
a \code{matrix} that contains normalized data,
a \code{matrix} that contains normalized initial direct-relation matrix,
a \code{data.frame} that contains relationships between criteria,
a \code{graph},
a \code{num} that contains threshold value,
a \code{list} of criteria comparisons.
}
\description{
Executes all functions and conducts dematel analysis at once
}
