% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitaldata.R
\docType{data}
\name{hospitaldata}
\alias{hospitaldata}
\title{Hospital Location Selection Data}
\format{
A data frame of 10 rows and 10 columns
\describe{
 \item{K1}{Numeric values of land / building cost}
 \item{K2}{Numeric values of operational repair / maintenance costs}
 \item{K3}{Numeric values of population densitity}
 \item{K4}{Numeric values of distance to educational institutions in the neighbourhood}
 \item{K5}{Numeric values of people's income level in the neighbourhood}
 \item{K6}{Numeric values of distance to public transport vehicles}
 \item{K7}{Numeric values of distance to suppliers}
 \item{K8}{Numeric values of distance to other hospitals}
 \item{K9}{Numeric values of building / land opportunities for additional units planned to be built in the future}
 \item{K10}{Numeric values of convenient transportation for ambulances}
}
}
\usage{
hospitaldata
}
\description{
A dataset containing hospital location selection problem data.
This data set gathered from Celikbilek Y., Ozdemir M. Cok Kriterli Karar Verme
Teknikleri Aciklamali ve Karsilastirmali Saglik Bilimleri Uygulamalari ile,
pp. 12, NOBEL Akademik Yayincilik, Ankara, 2020.
}
\keyword{datasets}
