% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data}
\usage{
normalize_data(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a \code{list} of data, and normalized matrix.
}
\description{
Normalizes matrix format data
}
