% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships_between_criteria.R
\name{relationships_between_criteria}
\alias{relationships_between_criteria}
\title{Relationships between criteria}
\usage{
relationships_between_criteria(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a \code{data.frame}
}
\description{
Returns total relationships between criteria data.frame
}
\examples{
relationships_between_criteria(dematel::hospitaldata)
relationships_between_criteria(dematel::nurseselection)
relationships_between_criteria(dematel::medicaldevice)
}
\seealso{
\link[base]{apply}
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
