% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_value.R
\name{threshold_value}
\alias{threshold_value}
\title{Threshold value}
\usage{
threshold_value(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a \code{num}
}
\description{
Returns threshold value of direct relationship decision matrix
}
\examples{
threshold_value(dematel::hospitaldata)
threshold_value(dematel::nurseselection)
threshold_value(dematel::medicaldevice)
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
