% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_relationship_matrix.R
\name{total_relationship_matrix}
\alias{total_relationship_matrix}
\title{Relationship matrix}
\usage{
total_relationship_matrix(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a \code{matrix}
}
\description{
Returns total relationship matrix of direct relationship decision matrix
}
\examples{
total_relationship_matrix(dematel::hospitaldata)
total_relationship_matrix(dematel::nurseselection)
total_relationship_matrix(dematel::medicaldevice)
}
\seealso{
\code{\link{apply}} function.
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
