% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Causal Diagram}
\usage{
visualize(x, data_control = TRUE)
}
\arguments{
\item{x}{a matrix containing the values of direct relationship decision matrix.}

\item{data_control}{is a pre-defined logical parameter that whether data should checked.}
}
\value{
This function returns a graph
}
\description{
Returns Causal Diagram of criteria
}
\examples{
visualize(dematel::hospitaldata)
visualize(dematel::nurseselection)
visualize(dematel::medicaldevice)
}
\author{
Muhlis Ozdemir <muhlisoz@gmail.com>
}
